exchange/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alpha-vantage==2.3.1
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
Django==4.0
django-environ==0.8.1
djangorestframework==3.13.1
frozenlist==1.2.0
idna==3.3
multidict==5.2.0
psycopg2==2.9.2
pytz==2021.3
requests==2.26.0
schedule==1.1.0
sqlparse==0.4.2
typing-extensions==4.0.1
urllib3==1.26.7
yarl==1.7.2
